requirements.txt
numpy>=1.23.5,<2.0.0
pandas>=2.2.3
matplotlib>=3.10.1
tensorflow>=2.15.0
scikit-learn>=1.3.0
seaborn>=0.13.0
joblib>=1.3.0
pyyaml>=6.0.1
pyarrow>=19.0.1
